requirements.txt
pygame-ce
pygbag
numpy
pillow
